🎨 Refactored component file structure and import paths, Improved dark mode colors
