+++ src/components/ui/Button/mode-toggle.tsx
@@ -1,10 +1,10 @@
-import { Button } from "@/components/ui/button"
+import { Button } from "@/components/ui/Button/button"
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu"
+} from "@/components/ui/Dropdown/dropdown-menu"
 import useTheme from "@/hooks/useSelectTheme"
 import { Moon, Sun } from "lucide-react"
