--- src/config/site.ts
@@ -1,18 +1,0 @@
-export type SiteConfig = typeof siteConfig
-
-export const siteConfig = {
-  name: "Vite",
-  description:
-    "Beautifully designed components built with Radix UI and Tailwind CSS.",
-  mainNav: [
-    {
-      title: "Home",
-      href: "/",
-    },
-  ],
-  links: {
-    youtube: "https://youtube.com/@m6io",
-    github: "https://github.com/m6io/shadcn-vite-template",
-    docs: "https://ui.shadcn.com",
-  },
-}
