+++ src/index.css
@@ -1,7 +1,7 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
- 
+
 @layer base {
   :root {
     --background: 0 0% 100%;
@@ -9,63 +9,64 @@
 
     --card: 0 0% 100%;
     --card-foreground: 222.2 84% 4.9%;
- 
+
     --popover: 0 0% 100%;
     --popover-foreground: 222.2 84% 4.9%;
- 
+
     --primary: 222.2 47.4% 11.2%;
     --primary-foreground: 210 40% 98%;
- 
+
     --secondary: 210 40% 96.1%;
     --secondary-foreground: 222.2 47.4% 11.2%;
- 
+
     --muted: 210 40% 96.1%;
     --muted-foreground: 215.4 16.3% 46.9%;
- 
+
     --accent: 210 40% 96.1%;
     --accent-foreground: 222.2 47.4% 11.2%;
- 
+
     --destructive: 0 84.2% 60.2%;
     --destructive-foreground: 210 40% 98%;
 
     --border: 214.3 31.8% 91.4%;
     --input: 214.3 31.8% 91.4%;
     --ring: 222.2 84% 4.9%;
- 
+
     --radius: 0.5rem;
   }
+
+.dark {
+    --background: 0 0% 10%; /* Lighter black, near dark gray */
+    --foreground: 0 0% 90%; /* Light gray for text on dark background */
  
-  .dark {
-    --background: 222.2 84% 4.9%;
-    --foreground: 210 40% 98%;
+    --card: 0 0% 12%; /* Slightly lighter black for cards */
+    --card-foreground: 0 0% 90%; /* Light gray for card text */
  
-    --card: 222.2 84% 4.9%;
-    --card-foreground: 210 40% 98%;
+    --popover: 0 0% 12%; /* Lighter black for popovers */
+    --popover-foreground: 0 0% 90%; /* Light gray for popover text */
  
-    --popover: 222.2 84% 4.9%;
-    --popover-foreground: 210 40% 98%;
+    --primary: 0 0% 90%; /* Light gray for primary text */
+    --primary-foreground: 0 0% 15%; /* Darker gray for primary foreground */
  
-    --primary: 210 40% 98%;
-    --primary-foreground: 222.2 47.4% 11.2%;
+    --secondary: 0 0% 25%; /* Medium gray for secondary elements */
+    --secondary-foreground: 0 0% 90%; /* Light gray for secondary text */
  
-    --secondary: 217.2 32.6% 17.5%;
-    --secondary-foreground: 210 40% 98%;
+    --muted: 0 0% 25%; /* Muted gray tone */
+    --muted-foreground: 0 0% 70%; /* Lighter gray for muted foreground */
  
-    --muted: 217.2 32.6% 17.5%;
-    --muted-foreground: 215 20.2% 65.1%;
+    --accent: 0 0% 25%; /* Gray accent color */
+    --accent-foreground: 0 0% 90%; /* Light gray for accent foreground */
  
-    --accent: 217.2 32.6% 17.5%;
-    --accent-foreground: 210 40% 98%;
+    --destructive: 0 60% 40%; /* A slightly more vibrant red for destructive actions */
+    --destructive-foreground: 0 0% 90%; /* Light gray for destructive foreground */
  
-    --destructive: 0 62.8% 30.6%;
-    --destructive-foreground: 210 40% 98%;
- 
-    --border: 217.2 32.6% 17.5%;
-    --input: 217.2 32.6% 17.5%;
-    --ring: 212.7 26.8% 83.9%;
-  }
+    --border: 0 0% 25%; /* Gray for borders */
+    --input: 0 0% 25%; /* Gray for input background */
+    --ring: 0 0% 50%; /* Light gray for ring/focus effect */
 }
- 
+
+}
+
 @layer base {
   * {
     @apply border-border;
